Swap Display Name with Device Network ID
This is just a simple swap between using the name to denote who and where the presence is for from the Display Name/Display Label (what most people organize by) to the Device Network ID. This allows for the user to change the display name/label as they please for organization purposes of their devices without breaking any functionality. It would require a modification of the setup-guide however.